Move `GTATranslateShiftKey` and `GTATranslateKey` to `WinPlatform.h`
